examples/python/requirements.txt
panini-lang>=0.1.0
python-dotenv
openai
anthropic
google-generativeai
